refactor(embedding): extract csm-embedding into standalone workspace crate
- Create crates/csm-embedding/ with EmbeddingProvider trait and providers
- Move embedding modules from src/embedding/ to csm-embedding crate
- Update main crate to re-export from csm-embedding
- Add csm-embedding to workspace members
- Fix rand::RngExt import in csm-core bundle_simd tests

Closes #364
