+++ crates/csm-embedding/src/projection.rs
