+++ llms-full.txt
@@ -17,6 +17,7 @@
 - cloudevents-sdk (0.9.0)
 - colored (2.2.0)
 - csm-core
+- csm-embedding
 - fastembed (4)
 - glob (0.3.2)
 - hnsw_rs (0.3.4)
@@ -38,9 +39,9 @@
 - cli: [dep:clap, dep:clap_complete, dep:anyhow, dep:colored, dep:glob]
 - cloudevents: [dep:cloudevents-sdk, dep:uuid, uuid/js]
 - default: [cli, persistence, parallel]
-- embed-fastembed: [dep:fastembed]
-- embed-openai: [dep:reqwest]
-- embed-voyage: [dep:reqwest]
+- embed-fastembed: [csm-embedding/embed-fastembed]
+- embed-openai: [csm-embedding/embed-openai]
+- embed-voyage: [csm-embedding/embed-voyage]
 - events-http: [cloudevents, cloudevents-sdk/http-binding, dep:reqwest]
 - mcp: [dep:rmcp, cli]
 - otlp-json: []
@@ -311,47 +312,9 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub struct ConceptBuilder
 - impl ConceptBuilder
 
-### src/embedding/fastembed.rs
-
-- pub struct FastEmbedProvider
-- impl FastEmbedProvider
-- impl EmbeddingProvider for FastEmbedProvider
-
-### src/embedding/hdc_text.rs
-
-- pub struct HdcTextProvider
-- impl Default for HdcTextProvider
-- impl HdcTextProvider
-- impl EmbeddingProvider for HdcTextProvider
-
 ### src/embedding/mod.rs
 
-- pub use hdc_text::HdcTextProvider
-- pub use projection::{Projection, ProjectionConfig}
-- pub use fastembed::FastEmbedProvider
-- pub use remote_openai::OpenAiProvider
-- pub use remote_voyage::VoyageProvider
-- pub trait EmbeddingProvider
-- pub fn get_provider
-
-### src/embedding/projection.rs
-
-- pub struct ProjectionConfig
-- impl Default for ProjectionConfig
-- pub struct Projection
-- impl Projection
-
-### src/embedding/remote_openai.rs
-
-- pub struct OpenAiProvider
-- impl OpenAiProvider
-- impl EmbeddingProvider for OpenAiProvider
-
-### src/embedding/remote_voyage.rs
-
-- pub struct VoyageProvider
-- impl VoyageProvider
-- impl EmbeddingProvider for VoyageProvider
+- pub use csm_embedding::*
 
 ### src/export_payload.rs
 
@@ -2523,252 +2486,12 @@ impl ConceptBuilder {
 }
 ```
 
-## src/embedding/fastembed.rs
-
-### FastEmbedProvider
-
-```rust
-pub struct FastEmbedProvider {
-}
-```
-
-FastEmbed local embedding provider.
-
-Uses ONNX runtime with downloaded models. No API key required.
-Models are cached locally after first download.
-
-Note: Does not implement Debug because fastembed::TextEmbedding doesn't.
-
-### impl EmbeddingProvider for FastEmbedProvider
-
-```rust
-impl EmbeddingProvider for FastEmbedProvider {
-}
-```
-
-### impl FastEmbedProvider
-
-```rust
-impl FastEmbedProvider {
-    pub fn new() -> Result<Self>;
-    pub fn with_model(model_name: &str) -> Result<Self>;
-    pub fn model_name(&self) -> &str;
-}
-```
-
-## src/embedding/hdc_text.rs
-
-### HdcTextProvider
-
-```rust
-#[derive(Debug, Clone)]
-pub struct HdcTextProvider {
-}
-```
-
-HDC text embedding provider using FNV-1a hashing.
-
-This is the default provider when no external embeddings are configured.
-It generates hypervectors directly without an intermediate f32 embedding,
-so the "native dimension" is 10240 and projection is identity.
-
-### impl Default for HdcTextProvider
-
-```rust
-impl Default for HdcTextProvider {
-}
-```
-
-### impl EmbeddingProvider for HdcTextProvider
-
-```rust
-impl EmbeddingProvider for HdcTextProvider {
-}
-```
-
-### impl HdcTextProvider
-
-```rust
-impl HdcTextProvider {
-    pub fn new() -> Self;
-    pub fn with_config(config: csm_core::encoder::TextEncoderConfig) -> Self;
-}
-```
-
 ## src/embedding/mod.rs
 
-### EmbeddingProvider
+### csm_embedding::*
 
 ```rust
-pub trait EmbeddingProvider: Send + Sync {
-    fn name(&self) -> &str;
-    fn native_dim(&self) -> usize;
-    fn embed(&self, text: &str) -> Result<Vec<f32>>;
-    fn embed_batch(&self, texts: &[&str]) -> Result<Vec<Vec<f32>>>;
-    fn project(&self, vec: &[f32], projection: &Projection) -> HVec10240;
-}
-```
-
-Embedding provider trait for text-to-vector conversion.
-
-Implementations may be:
-- Local (HDC hash, FastEmbed ONNX)
-- Remote HTTP (OpenAI, Voyage)
-
-All providers project their native dimensionality to HVec10240.
-
-### fastembed::FastEmbedProvider
-
-```rust
-pub use fastembed::FastEmbedProvider;
-```
-
-### get_provider
-
-```rust
-pub fn get_provider(name: &str) -> Result<std::sync::Arc<dyn Trait>>
-```
-
-Factory to get an embedding provider by name.
-
-Format: "provider_name" or "provider_name:model_name"
-
-### hdc_text::HdcTextProvider
-
-```rust
-pub use hdc_text::HdcTextProvider;
-```
-
-### projection::{Projection, ProjectionConfig}
-
-```rust
-pub use projection::{Projection, ProjectionConfig};
-```
-
-### remote_openai::OpenAiProvider
-
-```rust
-pub use remote_openai::OpenAiProvider;
-```
-
-### remote_voyage::VoyageProvider
-
-```rust
-pub use remote_voyage::VoyageProvider;
-```
-
-## src/embedding/projection.rs
-
-### Projection
-
-```rust
-#[derive(Debug, Clone)]
-pub struct Projection {
-}
-```
-
-Sparse random projection matrix for f32[native_dim] → HVec10240.
-
-Matrix is 10240 × native_dim with values in {-1, 0, +1}.
-Stored as flat array with indices for non-zero entries.
-
-### ProjectionConfig
-
-```rust
-#[derive(Debug, Clone)]
-pub struct ProjectionConfig {
-    pub seed: u64,
-    pub native_dim: usize,
-    pub target_dim: usize,
-    pub sparsity: f32,
-}
-```
-
-Configuration for projection matrix generation.
-
-### impl Default for ProjectionConfig
-
-```rust
-impl Default for ProjectionConfig {
-}
-```
-
-### impl Projection
-
-```rust
-impl Projection {
-    pub fn empty() -> Self;
-    pub fn new(config: &ProjectionConfig) -> Self;
-    pub fn project(&self, vec: &[f32]) -> HVec10240;
-    pub fn nnz(&self) -> usize;
-    pub fn sparsity_ratio(&self) -> f32;
-}
-```
-
-## src/embedding/remote_openai.rs
-
-### OpenAiProvider
-
-```rust
-#[derive(Debug)]
-pub struct OpenAiProvider {
-}
-```
-
-OpenAI embedding provider via HTTP API.
-
-Default model: text-embedding-3-small (1536 dimensions).
-API key must be set via environment or constructor.
-
-### impl EmbeddingProvider for OpenAiProvider
-
-```rust
-impl EmbeddingProvider for OpenAiProvider {
-}
-```
-
-### impl OpenAiProvider
-
-```rust
-impl OpenAiProvider {
-    pub fn from_env() -> Result<Self>;
-    pub fn new(api_key: String) -> Result<Self>;
-    pub fn with_model(self, model: impl Trait) -> Self;
-    pub fn with_base_url(self, url: impl Trait) -> Self;
-}
-```
-
-## src/embedding/remote_voyage.rs
-
-### VoyageProvider
-
-```rust
-#[derive(Debug)]
-pub struct VoyageProvider {
-}
-```
-
-Voyage embedding provider via HTTP API.
-
-Default model: voyage-2 (1024 dimensions).
-API key must be set via environment or constructor.
-
-### impl EmbeddingProvider for VoyageProvider
-
-```rust
-impl EmbeddingProvider for VoyageProvider {
-}
-```
-
-### impl VoyageProvider
-
-```rust
-impl VoyageProvider {
-    pub fn from_env() -> Result<Self>;
-    pub fn new(api_key: String) -> Result<Self>;
-    pub fn with_model(self, model: impl Trait) -> Self;
-    pub fn with_base_url(self, url: impl Trait) -> Self;
-}
+pub use csm_embedding::*;
 ```
 
 ## src/export_payload.rs
