+++ opencode.json
@@ -1,9 +1,12 @@
 {
   "$schema": "https://opencode.ai/config.json",
+  "indexing": {
+    "enabled": true
+  },
   "provider": {
     "bailian-coding-plan": {
-      "npm": "@ai-sdk/anthropic",
       "name": "Model Studio Coding Plan",
+      "npm": "@ai-sdk/anthropic",
       "options": {
         "baseURL": "https://coding-intl.dashscope.aliyuncs.com/apps/anthropic/v1"
       },
@@ -12,62 +15,170 @@
           "name": "Qwen3.5 Plus",
           "reasoning": true,
           "tool_call": true,
-          "modalities": { "input": ["text", "image", "video"], "output": ["text"] },
-          "limit": { "context": 1000000, "output": 65536 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 1000000,
+            "output": 65536
+          },
+          "modalities": {
+            "input": [
+              "text",
+              "image",
+              "video"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "qwen3-max": {
           "name": "Qwen3 Max",
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 262144, "output": 65536 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 262144,
+            "output": 65536
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "qwen3-coder-next": {
           "name": "Qwen3 Coder Next",
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 262144, "output": 65536 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 262144,
+            "output": 65536
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "qwen3-coder-plus": {
           "name": "Qwen3 Coder Plus",
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 1048576, "output": 65536 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 1048576,
+            "output": 65536
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "glm-5": {
           "name": "GLM-5",
           "reasoning": true,
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 202752, "output": 16384 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 202752,
+            "output": 16384
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "glm-4.7": {
           "name": "GLM-4.7",
           "reasoning": true,
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 202752, "output": 16384 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 202752,
+            "output": 16384
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "minimax-m2.5": {
           "name": "MiniMax M2.5",
           "reasoning": true,
           "tool_call": true,
-          "modalities": { "input": ["text"], "output": ["text"] },
-          "limit": { "context": 1048576, "output": 32768 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 1048576,
+            "output": 32768
+          },
+          "modalities": {
+            "input": [
+              "text"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         },
         "kimi-k2.5": {
           "name": "Kimi K2.5",
           "reasoning": true,
           "tool_call": true,
-          "modalities": { "input": ["text", "image", "video"], "output": ["text"] },
-          "limit": { "context": 262144, "output": 32768 },
-          "cost": { "input": 0, "output": 0 }
+          "cost": {
+            "input": 0,
+            "output": 0
+          },
+          "limit": {
+            "context": 262144,
+            "output": 32768
+          },
+          "modalities": {
+            "input": [
+              "text",
+              "image",
+              "video"
+            ],
+            "output": [
+              "text"
+            ]
+          }
         }
       }
     }
